### 2019-01-28 #1212 @arfon
@arsilva87 - please review the submission criteria and verify that your submission meets these criteria: https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements

On first inspection this submission seems to be lacking thorough documentation, tests, or an open source license.
