### 2019-01-28 #1212 @arsilva87
Hello! Licence added to the software repository. Thank you!

Em seg, 28 de jan de 2019 às 10:19, Arfon Smith <notifications@github.com>
escreveu:

> @arsilva87 <https://github.com/arsilva87> - please review the submission
> criteria and verify that your submission meets these criteria:
> https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements
>
> On first inspection this submission seems to be lacking thorough
> documentation, tests, or an open source license.
>
