+++ DESCRIPTION
@@ -2,8 +2,8 @@ Package: biotools
 Type: Package
 Title: Tools for Biometry and Applied Statistics in Agricultural
         Science
-Version: 3.1
-Date: 2017-05-25
+Version: 5.0
+Date: 2019-01-29
 LazyLoad: yes
 LazyData: yes
 Author: Anderson Rodrigo da Silva
@@ -18,4 +18,4 @@ Description: Tools designed to perform and work with cluster analysis (including
 	power is implemented. biotools also contains the new tests for genetic covariance components.
 	An approach for predicting spatial gene diversity is implemented.
 License: GPL (>= 2)
-Built: R 3.2.1; ; 2017-05-25 16:15:19 UTC; windows
+Built: R 3.4.3; ; 2019-01-29 00:19:01 UTC; windows
