### 2019-04-04 #1321 @karthik
@arsilva87 Following up from the issue I raised in your repo, the reviewers and I are unable to install your software with reasonable effort. It's because of your dependency on rpanel, which requires a dated Tcltk widget called `BWidget`. Can you please articulate why you need these dependencies, how you are using them and why you are unable to refactor with modern alternatives? I'm pausing the review until we are able to resolve this.
