### 2019-04-05 #1321 @arsilva87
@karthik The problem was solved by removing the dependencies with none or minimal losses, only for a single functionality of the package.
