### 2019-04-29 #1321 @diyadas
The README still directs users to install from CRAN, which I gather has not been updated. Installation proceeds with `devtools::install_github` - please update the README.
