### 2019-04-29 #1321 @arsilva87
@diyadas README file updated.
