### 2019-05-10 #1321 @labarba
@diyadas — Could you check on the author improvements, continue with your checklist, and report back to us with anything missing? Thank you!
