### 2019-05-16 arsilva87/biotools#2 @diyadas
Clarify statement of need in README
As per openjournals/joss-reviews#1321.

paper.md contains some information about what problems the software is designed to solve, but it's not obvious from the README. This paragraph answers the question:
> There are several R packages for cluster analysis, with different outcomes and objective functions. However, none of them offers an implementation of the Tocher's algorithm. The R package biotools contains an implementation of Tocher's algorithms, the original and the modified or "sequential" algorithm [@Vasconcelos07], as well as tools for evaluating the quality of clustering outcome. For this last part, biotools supplies some new and standard techniques such as: cophenetic() (for class tocher) - a specific cophenetic correlation coefficient [@Silva13], boxM() - the Box's M-test for evaluating the equality of the cluster covariance matrices and D2.disc() - a discriminant analysis based on Mahalanobis distances. In addition, the function singh() can be used for determining the importance of variables based on the squared generalized Mahalanobis distance [@Singh81].

Please also specify who the target audience is in the README
