### 2019-05-16 arsilva87/biotools#3 @diyadas
Clarify example usage
As per openjournals/joss-reviews#1321

Example code is given in the README, but it would be helpful if you added comments / annotated it in some way (what is the "real-world analysis problem"?)
