### 2019-05-16 arsilva87/biotools#5 @diyadas
Please add community guidelines
As per openjournals/joss-reviews#1321

I was not able to find community guidelines. As per the reviewing rubric:
>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The `usethis` package has a function `use_tidy_contributing` that can guide you through this.
