### 2019-05-16 #1321 @diyadas
Thanks for your patience. I've opened issues in your repo @arsilva87. I'll be happy to take a look again after they're addressed.
