+++ README.md
@@ -1,6 +1,54 @@
 # biotools
 
-Tools designed to perform and work with cluster analysis (including Tocher's algorithm), discriminant analysis and path analysis (standard and under collinearity), as well as some useful miscellaneous tools for dealing with sample size and optimum plot size calculations. Mantel's permutation test can be found in this package. A new approach for calculating its power is implemented. biotools also contains the new tests for genetic covariance components. An approach for predicting spatial gene diversity is implemented.
+Tools designed to perform and work with cluster analysis (including Tocher's algorithm), discriminant analysis and path analysis (standard and under collinearity), as well as some useful miscellaneous tools for dealing with sample size and optimum plot size calculations. Mantel's permutation test can be found in this package. A new approach for calculating its power is implemented. biotools also contains the new tests for genetic covariance components. An approach for predicting spatial gene diversity is implemented. Some of the main tools are described here.
+
+# Importance of variables: Singh's criterion
+
+We have implemented in the function singh() the method proposed by Singh (1981) for determining the importance of variables based on the squared generalized Mahalanobis distance. In his approach, the importance of the $j$-th variable ($j = 1, 2, ..., p$) on the calculation of the distance matrix can be obtained by:
+
+\begin{equation}
+	S_{.j} = \sum_{i=1}^{n-1} \sum_{i'>i}^{n} (x_{ij} - x_{i'j}) ({\bf x}_i - {\bf x}_{i'})^T \mat{\Sigma}_{.j}^{-1} 
+\end{equation}
+
+\noindent where $x_{ij}$ and $x_{i'j}$ are the observation taken at the $i$-th and $i'$-th objects (individuals) for the $j$-th variable, ${\bf x}_i$ is the $p$-variate vector of the $i$-th object and $\mat{\Sigma}_{.j}^{-1}$ is the $j$-th column of the inverse of the covariance matrix.
+
+The output of singh() is basically a matrix containing every $S_{.j}$ and its relative value.
+
+# Tocher's algorithm
+
+biotools contains the method suggested by K.D. Tocher (Rao 1952) for clustering objects, which consists of an optimization method that follows the algorithm:
+
+\begin{enumerate}
+	\item (Input) Take a distance matrix, let us say ${\bf d}$, of size $n$.
+	\item Define a clustering criterion, i.e., a limit of distance at which to evaluate the inclusion of objects in a cluster in formation. Usually, this criterion is defined as follows: let ${\bf m}$ be a vector of size $n$ containing the lowest distances involving each object, then take $\theta = \max{({\bf m})}$ to be the clustering criterion.
+	\item Locate the pair of objects with the lowest distance in ${\bf d}$ to be the starting cluster.
+	\item Compute the average distance ($d_{k(j)}$) of the starting cluster, say $k$, to a certain object, say $j$, that shows the lowest pairwise distance with each object in the starting cluster.
+	\item Compare the statistic evaluated at the step 4 with $\theta$. Then, add the new object to the starting cluster if $d_{k(j)} \leq \theta$. Otherwise, go to step 3 not considering the objects already clustered and start a new cluster.
+\end{enumerate}
+
+The process continues until the last remaining object is evaluated and either included in the last cluster formed or allocated to its own separated cluster. The function {\ttfamily tocher()} performs optimization clustering and returns an object of class 'tocher', which contains the clusters formed, a numeric vector indicating the cluster of each object, a matrix of cluster distances and also the input, a 'dist' object.
+
+# Cluster distances
+
[…]
