+++ README.md
@@ -6,7 +6,7 @@ Tools designed to perform and work with cluster analysis (including Tocher's alg
 
 We have implemented in the function singh() the method proposed by Singh (1981) for determining the importance of variables based on the squared generalized Mahalanobis distance. In his approach, the importance of the $j$-th variable ($j = 1, 2, ..., p$) on the calculation of the distance matrix can be obtained by:
 
-Let $\sum_{i=1}$
+![](https://latex.codecogs.com/gif.latex?S_%7B.j%7D%20%3D%20%5Csum_%7Bi%3D1%7D%5E%7Bn-1%7D%20%5Csum_%7Bi%27%3Ei%7D%5E%7Bn%7D%20%28x_%7Bij%7D%20-%20x_%7Bi%27j%7D%29%20%28%7B%5Cbf%20x%7D_i%20-%20%7B%5Cbf%20x%7D_%7Bi%27%7D%29%5ET%20%5Cmat%7B%5CSigma%7D_%7B.j%7D%5E%7B-1%7D)
 
 \begin{equation}
 	S_{.j} = \sum_{i=1}^{n-1} \sum_{i'>i}^{n} (x_{ij} - x_{i'j}) ({\bf x}_i - {\bf x}_{i'})^T \mat{\Sigma}_{.j}^{-1} 
