+++ README.md
@@ -4,51 +4,43 @@ Tools designed to perform and work with cluster analysis (including Tocher's alg
 
 # Importance of variables: Singh's criterion
 
-We have implemented in the function singh() the method proposed by Singh (1981) for determining the importance of variables based on the squared generalized Mahalanobis distance. In his approach, the importance of the $j$-th variable ($j = 1, 2, ..., p$) on the calculation of the distance matrix can be obtained by:
+The function singh() runs the method proposed by Singh (1981) for determining the importance of variables based on the squared generalized Mahalanobis distance. In his approach, the importance of the j-th variable (j = 1, 2, ..., p) on the calculation of the distance matrix can be obtained by:
 
 ![](https://latex.codecogs.com/gif.latex?S_%7B.j%7D%20%3D%20%5Csum_%7Bi%3D1%7D%5E%7Bn-1%7D%20%5Csum_%7Bi%27%3Ei%7D%5E%7Bn%7D%20%28x_%7Bij%7D%20-%20x_%7Bi%27j%7D%29%20%28%7B%5Cbf%20x%7D_i%20-%20%7B%5Cbf%20x%7D_%7Bi%27%7D%29%5ET%20%5Cmat%7B%5CSigma%7D_%7B.j%7D%5E%7B-1%7D)
 
-\begin{equation}
-	S_{.j} = \sum_{i=1}^{n-1} \sum_{i'>i}^{n} (x_{ij} - x_{i'j}) ({\bf x}_i - {\bf x}_{i'})^T \mat{\Sigma}_{.j}^{-1} 
-\end{equation}
+\noindent where ![](https://latex.codecogs.com/gif.latex?x_%7Bij%7D) and ![](https://latex.codecogs.com/gif.latex?x_%7Bi%27j%7D) are the observation taken at the i-th and i'-th objects (individuals) for the j-th variable, ![](https://latex.codecogs.com/gif.latex?%7B%5Cbf%20x%7D_i) is the p-variate vector of the i-th object and ![](https://latex.codecogs.com/gif.latex?%5Cmat%7B%5CSigma%7D_%7B.j%7D%5E%7B-1%7D) is the j-th column of the inverse of the covariance matrix.
 
-\noindent where $x_{ij}$ and $x_{i'j}$ are the observation taken at the $i$-th and $i'$-th objects (individuals) for the $j$-th variable, ${\bf x}_i$ is the $p$-variate vector of the $i$-th object and $\mat{\Sigma}_{.j}^{-1}$ is the $j$-th column of the inverse of the covariance matrix.
-
-The output of singh() is basically a matrix containing every $S_{.j}$ and its relative value.
+The output of singh() is basically a matrix containing every ![](https://latex.codecogs.com/gif.latex?S_%7B.j%7D) and its relative value.
 
 # Tocher's algorithm
 
 biotools contains the method suggested by K.D. Tocher (Rao 1952) for clustering objects, which consists of an optimization method that follows the algorithm:
 
-\begin{enumerate}
-	\item (Input) Take a distance matrix, let us say ${\bf d}$, of size $n$.
-	\item Define a clustering criterion, i.e., a limit of distance at which to evaluate the inclusion of objects in a cluster in formation. Usually, this criterion is defined as follows: let ${\bf m}$ be a vector of size $n$ containing the lowest distances involving each object, then take $\theta = \max{({\bf m})}$ to be the clustering criterion.
-	\item Locate the pair of objects with the lowest distance in ${\bf d}$ to be the starting cluster.
-	\item Compute the average distance ($d_{k(j)}$) of the starting cluster, say $k$, to a certain object, say $j$, that shows the lowest pairwise distance with each object in the starting cluster.
[…]
