+++ README.md
@@ -20,7 +20,7 @@ biotools contains the method suggested by K.D. Tocher (Rao 1952) for clustering 
 2. Define a clustering criterion, i.e., a limit of distance at which to evaluate the inclusion of objects in a cluster in formation. Usually, this criterion is defined as follows: ![](https://latex.codecogs.com/gif.latex?%7B%5Cbf%20m%7D) be a vector of size n containing the lowest distances involving each object, then take ![](https://latex.codecogs.com/gif.latex?%5Ctheta%20%3D%20%5Cmax%7B%28%7B%5Cbf%20m%7D%29%7D) to be the clustering criterion.
 3. Locate the pair of objects with the lowest distance in ![](https://latex.codecogs.com/gif.latex?%7B%5Cbf%20d%7D) to be the starting cluster.
 4. Compute the average distance ![](https://latex.codecogs.com/gif.latex?d_%7Bk%28j%29%7D) of the starting cluster, say k, to a certain object, say j, that shows the lowest pairwise distance with each object in the starting cluster.
-5. Compare the statistic evaluated at the step 4 with $\theta$. Then, add the new object to the starting cluster if ![](https://latex.codecogs.com/gif.latex?d_%7Bk%28j%29%7D%20%5Cleq%20%5Ctheta). Otherwise, go to step 3 not considering the objects already clustered and start a new cluster.
+5. Compare the statistic evaluated at the step 4 with ![](https://latex.codecogs.com/gif.latex?%5Ctheta). Then, add the new object to the starting cluster if ![](https://latex.codecogs.com/gif.latex?d_%7Bk%28j%29%7D%20%5Cleq%20%5Ctheta). Otherwise, go to step 3 not considering the objects already clustered and start a new cluster.
 
 The process continues until the last remaining object is evaluated and either included in the last cluster formed or allocated to its own separated cluster. The function tocher() performs optimization clustering and returns an object of class 'tocher', which contains the clusters formed, a numeric vector indicating the cluster of each object, a matrix of cluster distances and also the input, a 'dist' object.
 
