### 2019-06-08 arsilva87/biotools#2 @arsilva87
@diyadas @labarba the README file was updated/improved.
