+++ README.md
@@ -108,6 +108,6 @@ cor(cof, garlicdist)  # cophenetic correlation coefficient
 [1] 0.9086886
 ```
 
-# Report a bug
-Reach out directly by e-mail: anderson.silva@ifgoiano.edu.br.
+# Contributions and bug reports
 
+biotools is an ongoing project. Then, contributions are very welcome. If you have a question or have found a bug, please open an ![Issue](https://github.com/arsilva87/biotools/issues) or reach out directly by e-mail: anderson.silva@ifgoiano.edu.br.
