### 2019-06-09 #1321 @arsilva87
@labarba I have updated the repository and worked on the issued pointed out by @diyadas. Thanks.
