### 2019-06-25 98a818d
Delete test_tocher.R
