### 2019-06-25 759bd86
Delete testthat.R
