+++ tests/testthat/test_tocher.R
@@ -1,0 +1,24 @@
+library(testthat)
+context("test biotools")
+
+test_that("biotools works", {
+   data(iris)
+   mod <- manova(as.matrix(iris[, -5]) ~ Species, data = iris)
+   cov_r <- summary(mod)$SS$Residuals / mod$df.residual
+   s <- singh(iris[, -5], cov_r)
+      expect_output(str(s), "List of 2")
+   bm <- boxM(iris[, -5], iris[, 5])
+      expect_output(str(bm), "List of 8")
+   d2 <- D2.dist(iris[, -5], cov_r)
+      expect_output(str(d2), "Class 'dist'")
+   to <- tocher(d2)
+      expect_output(str(to), "List of 7")
+   dc <- distClust(d2, sapply(to$clusters, length), rownames(iris))
+      expect_output(str(dc), "num")
+   disc <- D2.disc(iris[, -5], to$class, cov_r)
+      expect_output(str(disc), "List of 6")
+   cm <- confusionmatrix(to$class, predict(disc)$class)
+      expect_output(str(cm), "num")
+   rate <- aer(to$class, predict(disc)$class)
+      expect_output(str(rate), "num")
+})
