### 2019-07-21 #1321 @arfon
@diyadas & @karthik, many thanks for your efforts thus far. As I have outlined in my email to @arsilva87, this submission still falls well below what is expected of a JOSS submission. As such I am rejecting this submission and invite the author to resubmit if they are able to bring the submission up to standard.
